Finished a mostly working options menu include video, sound, and control submenus
Also includes a rework of the options in general.
(methods are still needed to enable saving options back to the options.conf)

git-svn-id: https://svn.code.sf.net/p/openpirates/code@8 3faa6364-4847-409f-afa2-7fad3069fade

+++ src/global.cpp
@@ -82,7 +82,7 @@ int32_t a_to_i( const std::string& line )
     return number;
 }
 
-std::string i_to_a( int16_t num )
+std::string i_to_a( int32_t num )
 {
     std::string str;
     std::stringstream ss;
